Refactor: Rename webview_gtk to webkit2gtk_engine
